Criação dos menus diretor de curso e criação de curso no painel da instituição; Correção de bugs associados aos convites; Implementação de convite de diretor de curso (users que estejam ou não na plataforma); Correção de erro associado ao editar atividades; Correção de erro associado ao atualizar_turmas_ano_letivo; Editar curso criado (ainda faltam implementações, visualizar todo.md);  Criação do trata_login_convite para o utilizador que ja tem conta na plataforma mas não estar logado poder receber convite e funcionar; Remoção de curso da plataforma; Criação de alguns displays de erros a partir da session;

+++ modules/trata_convite_user_curso.mod.php
@@ -17,14 +17,49 @@ if(count($res) == 0) {
 }
 pr($res);
 $id_user = $res[0]['id'];
-$id_curso = $res[0]['id_curso'];
-$cargo = $res[0]['cargo'];
+if(isset($_SESSION['id'])) {
+    if($_SESSION['id'] == $id_user) {
+        $id_curso = $res[0]['id_curso'];
+        $cargo = $res[0]['cargo'];
+        if($cargo == 'Diretor de Curso') {
+            $cargo = 'Professor';
+            $sql = "UPDATE curso SET id_diretor_curso = $id_user WHERE id = $id_curso";
+            echo $sql;
+            my_query($sql);
+            $sql = "UPDATE curso SET ativo = 1 WHERE id = $id_curso";
+            my_query($sql);
+        } else {
+            $dc = false;
+        }
 
-$sql = "UPDATE rel_user_curso SET estado = '1' WHERE id_user = $id_user AND id_curso = $id_curso";
-$res = my_query($sql);
 
-$sql = "DELETE FROM conf_convite WHERE id = " . $_GET['convite'];
-$res = my_query($sql);
 
-$_SESSION['convite_aceite'] = true;
-header('Location:' . $arrConfig['url_admin'] . 'turma.php');
+        $sql = "UPDATE rel_user_curso SET estado = '1' WHERE id_user = $id_user AND id_curso = $id_curso";
+        $res = my_query($sql);
+
+        $sql = "DELETE FROM conf_convite WHERE id = " . $_GET['convite'];
+        $res = my_query($sql);
+
+        $_SESSION['convite_aceite'] = true;
+
+        header('Location:' . $arrConfig['url_admin'] . 'turma.php');
+    } else {
+        header('Location:' . $arrConfig['url_admin'] . 'index.php');
+    }
+
+} else {
+    if(count($res) > 0) {
+        $_SESSION['convite'] = $_GET['convite'];
+        $_SESSION['id_user'] = $res[0]['id'];
+        $_SESSION['email'] = $res[0]['email'];
+        $_SESSION['id_curso'] = $res[0]['id_curso'];
+        if($res[0]['cargo'] == 'Diretor de Curso') {            
+            $_SESSION['dc'] = true;
+        } else {
+            $_SESSION['dc'] = false;            
+        }
+        header('Location: ' . $arrConfig['url_paginas'] . 'auth/login.php' );
+    } else {
+        echo 'não foram encontrados dados para o convite em questão, podendo este ter sido retirado ou já ter sido utilizado.';
+    }
+}
